Add LaserPositionCard component and refactor cards
