+++ play-laser/app/(tabs)/workout/start/start-routine.tsx
@@ -2,12 +2,11 @@ import { StyleSheet, ScrollView } from 'react-native';
 
 import { View } from '@/components/Themed';
 import { PaperProvider, Text, Button, ActivityIndicator } from 'react-native-paper';
-import FontAwesome from '@expo/vector-icons/FontAwesome';
 import Colors from '@/constants/Colors';
 import { useColorScheme } from '@/components/useColorScheme';
-import { Workout, exampleWorkouts } from '@/types';
+import LaserPositionCard from '@/components/LaserPositionCard';
+import { Workout } from '@/types';
 import { router } from 'expo-router';
-import { LaserPositionCardProps, LaserGridProps } from '@/types';
 
 import React, { useState, useEffect } from 'react';
 import { fetchWorkouts } from "@/FirebaseConfig";
@@ -15,9 +14,7 @@ import { fetchWorkouts } from "@/FirebaseConfig";
 
 export default function WorkoutScreen() {
   const colorScheme = useColorScheme();
-  // const workouts: Workout[] = exampleWorkouts;
   const [workout, setWorkout] = useState<Workout | null>(null);
-  // console.log(workout);
 
   useEffect(() => {
     const workoutId = "1"; // TESTING BASIC 1 PREMADE ROUTINE
@@ -76,7 +73,7 @@ export default function WorkoutScreen() {
                 key={index}
                 workout={workout}
                 index={index}
-                laserPosition={index}
+                laserPosition={laserPosition}
               />
             ))}
           </ScrollView>
@@ -84,57 +81,6 @@ export default function WorkoutScreen() {
       </View>
     </PaperProvider>
   );
-}
-
-const LaserPositionCard: React.FC<LaserPositionCardProps> = ({ workout, laserPosition }) => {
-  const colorScheme = useColorScheme();
-
-  return (
-    <View style={[styles.laserPositionCard, {backgroundColor: Colors[colorScheme ?? 'light'].button}]}>
-      <LaserGrid 
-        numColumns={workout.numColumns}
-        numRows={workout.numRows} 
-        numPositions={workout.numPositions} 
-        laserPosition={workout.laserPositions[laserPosition]} 
-      />
-      <Text style={[styles.buttonText, {color: Colors[colorScheme ?? 'light'].buttonText}]}>
-        Laser {laserPosition + 1}
-      </Text>
-    </View>
-  )
-}
-
-const LaserGrid: React.FC<LaserGridProps> = ({ numColumns, numRows, numPositions, laserPosition }) => {
-  const colorScheme = useColorScheme();
-  const laserPositionRow = laserPosition != undefined ? Math.floor((laserPosition - 1) / numColumns) : -1;
-  const laserPositionColumn = laserPosition != undefined ? (laserPosition - 1) % numColumns : -1;
-
-  const rows = [];
-  for (let i = 0; i < numRows; i++) {
-    const columns = [];
-    for (let j = 0; j < numColumns; j++) {
-      columns.push(
-        <View key={`${i}-${j}`} style={[styles.gridItem, {backgroundColor: Colors[colorScheme ?? 'light'].button}]}>
-          <FontAwesome
-            name="dot-circle-o"
-            size={9}
-            color={(laserPositionRow == i && laserPositionColumn == j) ? "#422f7f" : "white"} 
-          />
-        </View>
-      );
-    }
-    rows.push(
-      <View key={i} style={[styles.gridRow, {backgroundColor: Colors[colorScheme ?? 'light'].button}]}>
-        {columns}
-      </View>
-    );
-  }
-
-  return (
-    <View style={[styles.laserGrid, {backgroundColor: Colors[colorScheme ?? 'light'].button}]}>
-      {rows}
-    </View>
-  )
 }
 
 const styles = StyleSheet.create({
@@ -175,29 +121,5 @@ const styles = StyleSheet.create({
     alignSelf: 'center',
     justifyContent: 'center',
     alignItems: 'center',
-  },
-  laserPositionCard: {
-    backgroundColor: 'white',
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'center',
-    padding: 8,
-    borderRadius: 16,
-    marginBottom: 16,
-  },
-  laserGrid: {
-    marginRight: 32,
-  },
-  gridRow: {
-    flexDirection: 'row',
-    justifyContent: 'center',
-    marginVertical: 5,
-  },
-  gridItem: {
-    width: 8,
-    height: 8,
-    alignItems: 'center',
-    justifyContent: 'center',
-    marginHorizontal: 5,
   },
 });
